fire spawning on buildings
Scripts updated
BuildingManager - holds 3 Lists for the fire, buildings, and data
InitiateBuildings - has access to this list in editor and prefabs or scriptable objects are drag/dropped there.
These objects are spawned to the scene.
The x,y points and number of windows for the fires are held in the scriptable objects.